Adding collectd package for Docker
This collectd build contains plugins for Riemann,
Python,Redis, Varnish.

I still want to add haproxy, mysql, postgresql,
and glusterfs.

I need to break this up into subpackages so it's not
so chunky at runtime.

This allows us to run docker containers for collectd that
monitor specific parts of our stack, as opposed to the
overall system metrics.